Rename fnmu to fnmi
The original function was nvmu which stood for "nvm upgrade". This made
sense since it reinstalled global packages from the current version to
the new version and was really only meant to be run when upgrading to
the latest Node version. Now, we always install a set list of global
packages since fnm doesn't support migrating global packages from the
current version, so "fnm install" makes more sense.